backend/requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-mysql==4.7.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
mysql-connector-python==8.0.30
mysqlclient==2.1.1
Pillow==9.2.0
protobuf==3.20.1
PyMySQL==1.0.2
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.2.4
tzdata==2022.2
